feat: add leave request submission with working day calculation, balance tracking, warnings and approver notification

+++ src/main/java/com/leavesync/workingday/WorkingDayResponse.java
@@ -1,12 +1,13 @@
 package com.leavesync.workingday;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record WorkingDayResponse(
 
         LocalDate from,
         LocalDate to,
-        int workingDays
+        BigDecimal workingDays
 
 ) {
 }
